@@ -1,6 +1,6 @@
 source 'https://rubygems.org'
 
-ruby '>=3.2.2'
+ruby '>=3.2.4'
 
 # Ensure github repositories are fetched using HTTPS
 git_source(:github) do |repo_name|
@@ -30,16 +29,16 @@ end
 # Optional libraries.  To conserve RAM, comment out any that you don't need,
 # then run `bundle` and commit the updated Gemfile and Gemfile.lock.
 gem 'erector', github: 'dsander/erector', branch: 'rails6'
-gem 'forecast_io', '~> 2.0.0'     # WeatherAgent
-gem 'hipchat', '~> 1.2.0'         # HipchatAgent
+gem 'hipchat', '~> 1.6.0' # HipchatAgent
 gem 'hypdf', bitbucket: 'knu/hypdf_gem', branch: 'uploadio_namespace' # PDFInfoAgent
-gem 'mini_racer'                  # JavaScriptAgent
-gem 'mqtt'                        # MQTTAgent
+gem 'mini_racer' # JavaScriptAgent
+gem 'mqtt' # MQTTAgent
 gem 'net-ftp'
-gem 'net-ftp-list'                # FtpsiteAgent
+gem 'net-ftp-list' # FtpsiteAgent
+gem 'pirate_weather_forecast_ruby' # WeatherAgent
 gem 'rturk', '~> 2.12.1'          # HumanTaskAgent
-gem 'slack-notifier', '~> 1.0.0'  # SlackAgent
-gem 'twilio-ruby', '~> 5.62.0'    # TwilioAgent
+gem 'slack-notifier', '~> 2.4.0'  # SlackAgent
+gem 'twilio-ruby', '~> 7.4.0'     # TwilioAgent
 gem 'xmpp4r', '~> 0.5.6'          # JabberAgent
 
 # Weibo Agents
@@ -90,47 +88,56 @@ end
 
 gem 'ace-rails-ap'
 gem 'bootsnap', require: false
-gem 'bootstrap-kaminari-views', '~> 0.0.3'
-gem 'bundler', '>= 1.5.0'
+gem 'bootstrap-kaminari-views', '~> 0.0.5'
+gem 'bundler', '>= 2.6.2'
 gem 'coffee-rails', '~> 5'
-gem 'daemons', '~> 1.1.9'
+gem 'daemons'
 gem 'delayed_job'
 gem 'delayed_job_active_record'
-gem 'devise', '~> 4.8'
-gem 'em-http-request', '~> 1.1.2'
+gem 'devise', '~> 4.9', '>= 4.9.4'
+gem 'em-http-request'
 gem 'execjs'
-gem 'faraday', '~> 1.0'
-gem 'faraday_middleware'
-gem 'feedjira', '~> 3.1'
-gem 'font-awesome-sass', '~> 4.7.0'
-gem 'foreman', '~> 0.87.2', github: 'ddollar/foreman'
-gem 'geokit', '~> 1.13'
-gem 'geokit-rails', '~> 2.3'
+gem 'faraday'
+# gem 'faraday-em_http'
+gem 'faraday-follow_redirects'
+gem 'faraday-gzip'
+# gem 'faraday-httpclient'
+gem 'faraday-typhoeus'
+gem 'feedjira', '~> 3.2', '>= 3.2.3'
+gem 'font-awesome-sass', '~> 6.7.2'
+gem 'foreman', '~> 0.88.1', github: 'ddollar/foreman'
+gem 'geokit', '~> 1.14'
+gem 'geokit-rails', '~> 2.5'
 gem 'httmultiparty', '~> 0.3.16'
-gem 'httparty', '~> 0.13'
+gem 'httparty', '~> 0.22'
 gem 'huginn_agent'
-gem 'jquery-rails', '~> 4.2.1'
-gem 'json', '~> 2.3'
-gem 'jsonpath', '~> 1.1'
-gem 'kaminari', '~> 1.2'
+gem 'jquery-rails', '~> 4.6'
+gem 'json', '~> 2.9', '>= 2.9.1'
+gem 'jsonpath', '~> 1.1', '>= 1.1.5'
+gem 'kaminari', '~> 1.2', '>= 1.2.2'
 gem 'kramdown'
-gem 'liquid', '~> 5.1'
-gem 'loofah', '~> 2.0'
+gem 'liquid', '~> 5.6'
+gem 'loofah', '~> 2.24'
 gem 'mail', '>= 2.8.1'
-gem 'mini_magick', ">= 4.9.4"
+gem 'mini_magick', '>= 5.0.1'
 gem 'multi_xml'
-gem "nokogiri", ">= 1.10.8"
+gem 'nokogiri', '>= 1.18.4'
 gem 'omniauth'
-gem 'rails', '~> 6.1.7'
-gem 'rails-html-sanitizer', '~> 1.2'
-gem 'rufus-scheduler', '~> 3.4', require: false
-gem 'sass-rails', '>= 6.0'
+gem 'rails', '~> 7.0.1'
+gem 'rails-html-sanitizer', '~> 1.6', '>= 1.6.2'
+gem 'rufus-scheduler', '~> 3.9', '>= 3.9.2', require: false
+gem 'sassc-rails'
 gem 'select2-rails'
 gem 'spectrum-rails'
 gem 'sprockets'
 gem 'terser'
-gem 'typhoeus', '~> 1.3.1'
-gem 'uglifier', '~> 2.7.2'
+gem 'typhoeus'
+gem 'uglifier'
+
+group :development, :test do
+  gem 'debug'
+  gem 'rspec-rails'
+end
 
 group :development do
   gem 'better_errors'
