@@ -1,8 +1,15 @@
 module Agents
   class GoogleTranslationAgent < Agent
     cannot_be_scheduled!
+    can_dry_run!
 
-    gem_dependency_check { defined?(Google) && defined?(Google::Cloud::Translate) }
+    gem_dependency_check do
+      require 'google/cloud/translate/v2'
+    rescue LoadError
+      false
+    else
+      true
+    end
 
     description <<~MD
       The Translation Agent will attempt to translate text between natural languages.
@@ -18,7 +25,10 @@ module Agents
 
       `from` is the language translated from. If it's not specified, the API will attempt to detect the source language automatically and return it within the response.
 
-      Specify what you would like to translate in `content` field, you can use [Liquid](https://github.com/huginn/huginn/wiki/Formatting-Events-using-Liquid) specify which part of the payload you want to translate.
+      Specify an object in `content` field using [Liquid](https://github.com/huginn/huginn/wiki/Formatting-Events-using-Liquid) expressions, which will be evaluated for each incoming event, and then translated to become the payload of the new event.
+      You can specify a nested object of any levels containing arrays and objects, and all string values except for object keys will be recursively translated.
+
+      Set `mode` to `merge` if you want to merge each translated content with the original event payload.  The default behavior (`clean`) is to emit events with only translated contents.
 
       `expected_receive_period_in_days` is the maximum number of days you would allow to pass between events.
     MD
@@ -27,13 +37,14 @@ module Agents
 
     def default_options
       {
-        'to' => "sv",
+        'mode' => 'clean',
+        'to' => 'sv',
         'from' => 'en',
         'google_api_key' => '',
         'expected_receive_period_in_days' => 1,
         'content' => {
           'text' => "{{message}}",
-          'moretext' => "{{another message}}"
+          'moretext' => "{{another_message}}"
         }
       }
     end
@@ -46,58 +57,83 @@ module Agents
       unless options['google_api_key'].present? && options['to'].present? && options['content'].present? && options['expected_receive_period_in_days'].present?
         errors.add :base, "google_api_key, to, content and expected_receive_period_in_days are all required"
       end
-    end
 
-    def translate_from
-      interpolated["from"].presence
+      case options['mode'].presence
+      when nil, /\A(?:clean|merge)\z|\{/
+        # ok
+      else
+        errors.add(:base, "mode must be 'clean' or 'merge'")
+      end
     end
 
     def receive(incoming_events)
       incoming_events.each do |event|
-        translated_event = {}
-        opts = interpolated(event)
-        opts['content'].each_pair do |key, value|
-          result = translate(value)
-          translated_event[key] = result.text
+        interpolate_with(event) do
+          translated_content = translate(interpolated['content'])
+
+          case interpolated['mode']
+          when 'merge'
+            create_event payload: event.payload.merge(translated_content)
+          else
+            create_event payload: translated_content
+          end
         end
-        create_event payload: translated_event
       end
     end
 
-    def google_client
-      @google_client ||= Google::APIClient.new(
-        {
-          application_name: "Huginn",
-          application_version: "0.0.1",
-          key: options['google_api_key'],
-          authorization: nil
-        }
-      )
-    end
+    def translate(content)
+      if !content.is_a?(Hash)
+        error("content must be an object, but it is #{content.class}.")
+        return
+      end
 
-    def translate_service
-      @translate_service ||= google_client.discovered_api('translate', 'v2')
-    end
-
-    def cloud_translate_service
-      # https://github.com/GoogleCloudPlatform/google-cloud-ruby/blob/master/google-cloud-translate/lib/google-cloud-translate.rb#L130
-      @google_client ||= Google::Cloud::Translate.new(
-        version: :v2,
+      api = Google::Cloud::Translate::V2.new(
         key: interpolated['google_api_key']
       )
-    end
 
-    def translate(value)
-      # google_client.execute(
-      #   api_method: translate_service.translations.list,
-      #   parameters: {
-      #     format: 'text',
-      #     source: translate_from,
-      #     target: options["to"],
-      #     q: value
-      #   }
-      # )
-      cloud_translate_service.translate(value, to: interpolated["to"], from: translate_from, format: "text")
+      texts = []
+      walker = ->(value) {
+        case value
+        in nil | Numeric | true | false
+        in _ if _.blank?
+        in String
+          texts << value
+        in Array
+          value.each(&walker)
+        in Hash
+          value.each_value(&walker)
+        end
+      }
+      walker.call(content)
+
+      translations =
+        if texts.empty?
+          []
+        else
+          api.translate(
+            *texts,
+            from: interpolated['from'].presence,
+            to: interpolated['to'],
+            format: 'text',
+          )
+        end
+
+      # Hash key order should be constant in Ruby
+      mapper = ->(value) {
+        case value
+        in nil | Numeric | true | false
+          value
+        in _ if _.blank?
+          value
+        in String
+          translations&.shift&.text
+        in Array
+          value.map(&mapper)
+        in Hash
+          value.transform_values(&mapper)
+        end
+      }
+      mapper.call(content)
     end
   end
 end
