@@ -30,7 +30,9 @@ module Agents
       * `sound` - the name of one of the sounds supported by device clients to override the user's default sound choice. [See PushOver docs for sound options.](https://pushover.net/api#sounds)
       * `retry` - Required for emergency priority - Specifies how often (in seconds) the Pushover servers will send the same notification to the user. Minimum value: `30`
       * `expire` - Required for emergency priority - Specifies how many seconds your notification will continue to be retried for (every retry seconds). Maximum value: `86400`
+      * `ttl` - set to a Time to Live in seconds
       * `html` - set to `true` to have Pushover's apps display the `message` content as HTML
+      * `monospace` - set to `true` to have Pushover's apps display the `message` content with a monospace font
 
     MD
 
@@ -48,7 +50,9 @@ module Agents
         'sound' => '{{ sound }}',
         'retry' => '{{ retry }}',
         'expire' => '{{ expire }}',
+        'ttl' => '{{ ttl }}',
         'html' => 'false',
+        'monospace' => 'false',
         'expected_receive_period_in_days' => '1'
       }
     end
@@ -86,6 +90,7 @@ module Agents
             sound
             retry
             expire
+            ttl
           ].each do |key|
             value = String.try_convert(interpolated[key].presence) or next
 
@@ -97,15 +102,21 @@ module Agents
             end
             post_params[key] = value
           end
-          # html is special because String.try_convert(true) gives nil (not even "nil", just nil)
-          if value = interpolated['html'].presence
-            post_params['html'] =
-              case value.to_s
-              when 'true', '1'
-                '1'
-              else
-                '0'
-              end
+
+          # boolean parameters
+          %w[
+            html
+            monospace
+          ].each do |key|
+            if value = interpolated[key].presence
+              post_params[key] =
+                case value.to_s
+                when 'true', '1'
+                  '1'
+                else
+                  '0'
+                end
+            end
           end
 
           send_notification(post_params)
