@@ -1,2 +1,4 @@
 require 'pp'
-HuginnAgent.require!
+Rails.application.config.to_prepare do
+  HuginnAgent.require!
+end
