@@ -3,7 +3,7 @@ require 'rails_helper'
 describe Agents::GoogleCalendarPublishAgent do
   let(:valid_params) do
     {
-      'expected_update_period_in_days' => "10",
+      'expected_update_period_in_days' => '10',
       'calendar_id' => calendar_id,
       'google' => {
         'key_file' => File.dirname(__FILE__) + '/../../data_fixtures/private.key',
@@ -14,7 +14,7 @@ describe Agents::GoogleCalendarPublishAgent do
   end
 
   let(:agent) do
-    _agent = Agents::GoogleCalendarPublishAgent.new(name: "somename", options: valid_params)
+    _agent = Agents::GoogleCalendarPublishAgent.new(name: 'somename', options: valid_params)
     _agent.user = users(:jane)
     _agent.save!
     _agent
@@ -24,8 +24,8 @@ describe Agents::GoogleCalendarPublishAgent do
     let(:message) do
       {
         'visibility' => 'default',
-        'summary' => "Awesome event",
-        'description' => "An example event with text. Pro tip: DateTimes are in RFC3339",
+        'summary' => 'Awesome event',
+        'description' => 'An example event with text. Pro tip: DateTimes are in RFC3339',
         'end' => {
           'date_time' => '2014-10-02T11:00:00-05:00'
         },
@@ -46,35 +46,36 @@ describe Agents::GoogleCalendarPublishAgent do
     let(:calendar_id) { 'sqv39gj35tc837gdns1g4d81cg@group.calendar.google.com' }
 
     let(:response_hash) do
-      {"kind"=>"calendar#event",
-        "etag"=>"\"2908684044040000\"",
-        "id"=>"baz",
-        "status"=>"confirmed",
-        "html_link"=>
-          "https://calendar.google.com/calendar/event?eid=foobar",
-        "created"=>"2016-02-01T15:53:41.000Z",
-        "updated"=>"2016-02-01T15:53:42.020Z",
-        "summary"=>"Awesome event",
-        "description"=>
-          "An example event with text. Pro tip: DateTimes are in RFC3339",
-        "creator"=>
-          {"email"=>
-            "blah-foobar@developer.gserviceaccount.com"},
-        "organizer"=>
-          {"email"=>calendar_id,
-            "display_name"=>"Huginn Location Log",
-            "self"=>true},
-        "start"=>{"date_time"=>"2014-10-03T00:30:00+09:30"},
-        "end"=>{"date_time"=>"2014-10-03T01:30:00+09:30"},
-        "i_cal_uid"=>"blah@google.com",
-        "sequence"=>0,
-        "reminders"=>{"use_default"=>true}
-      }
+      { 'kind' => 'calendar#event',
+        'etag' => '"2908684044040000"',
+        'id' => 'baz',
+        'status' => 'confirmed',
+        'html_link' =>
+          'https://calendar.google.com/calendar/event?eid=foobar',
+        'created' => '2016-02-01T15:53:41.000Z',
+        'updated' => '2016-02-01T15:53:42.020Z',
+        'summary' => 'Awesome event',
+        'description' =>
+          'An example event with text. Pro tip: DateTimes are in RFC3339',
+        'creator' =>
+          { 'email' =>
+            'blah-foobar@developer.gserviceaccount.com' },
+        'organizer' =>
+          { 'email' => calendar_id,
+            'display_name' => 'Huginn Location Log',
+            'self' => true },
+        'start' => { 'date_time' => '2014-10-03T00:30:00+09:30' },
+        'end' => { 'date_time' => '2014-10-03T01:30:00+09:30' },
+        'i_cal_uid' => 'blah@google.com',
+        'sequence' => 0,
+        'reminders' => { 'use_default' => true } }
     end
 
     def setup_mock!
-      fake_interface = Object.new
-      expect(GoogleCalendar).to receive(:new).with(agent.interpolate_options(agent.options), Rails.logger) { fake_interface }
+      fake_interface = double('fake_interface')
+      expect(GoogleCalendar).to receive(:new).with(agent.interpolate_options(agent.options), Rails.logger) {
+                                  fake_interface
+                                }
       expect(fake_interface).to receive(:publish_as).with(calendar_id, message) { response_hash }
       expect(fake_interface).to receive(:cleanup!)
     end
@@ -83,11 +84,12 @@ describe Agents::GoogleCalendarPublishAgent do
       it 'should publish any payload it receives' do
         setup_mock!
 
-        expect {
+        expect do
           agent.receive([event])
-        }.to change { agent.events.count }.by(1)
+        end.to change { agent.events.count }.by(1)
 
-        expect(agent.events.last.payload).to eq({ "success" => true, "published_calendar_event" => response_hash, "agent_id" => event.agent_id, "event_id" => event.id })
+        expect(agent.events.last.payload).to eq({ 'success' => true, 'published_calendar_event' => response_hash,
+                                                  'agent_id' => event.agent_id, 'event_id' => event.id })
       end
     end
 
@@ -104,7 +106,8 @@ describe Agents::GoogleCalendarPublishAgent do
         agent.receive([event])
 
         expect(agent.events.count).to eq(1)
-        expect(agent.events.last.payload).to eq({ "success" => true, "published_calendar_event" => response_hash, "agent_id" => event.agent_id, "event_id" => event.id })
+        expect(agent.events.last.payload).to eq({ 'success' => true, 'published_calendar_event' => response_hash,
+                                                  'agent_id' => event.agent_id, 'event_id' => event.id })
       end
 
       it 'should allow Liquid in the key' do
@@ -131,8 +134,8 @@ describe Agents::GoogleCalendarPublishAgent do
       _event.agent = agents(:bob_manual_event_agent)
       _event.payload = { 'message' => {
         'visibility' => 'default',
-        'summary' => "Awesome event",
-        'description' => "An example event with text. Pro tip: DateTimes are in RFC3339",
+        'summary' => 'Awesome event',
+        'description' => 'An example event with text. Pro tip: DateTimes are in RFC3339',
         'end' => {
           'dateTime' => '2014-10-02T11:00:00-05:00'
         },
@@ -148,8 +151,8 @@ describe Agents::GoogleCalendarPublishAgent do
     let(:message) do
       {
         'visibility' => 'default',
-        'summary' => "Awesome event",
-        'description' => "An example event with text. Pro tip: DateTimes are in RFC3339",
+        'summary' => 'Awesome event',
+        'description' => 'An example event with text. Pro tip: DateTimes are in RFC3339',
         'end' => {
           'date_time' => '2014-10-02T11:00:00-05:00'
         },
@@ -160,36 +163,37 @@ describe Agents::GoogleCalendarPublishAgent do
     end
 
     let(:response_hash) do
-      {"kind"=>"calendar#event",
-        "etag"=>"\"2908684044040000\"",
-        "id"=>"baz",
-        "status"=>"confirmed",
-        "html_link"=>
-          "https://calendar.google.com/calendar/event?eid=foobar",
-        "created"=>"2016-02-01T15:53:41.000Z",
-        "updated"=>"2016-02-01T15:53:42.020Z",
-        "summary"=>"Awesome event",
-        "description"=>
-          "An example event with text. Pro tip: DateTimes are in RFC3339",
-        "creator"=>
-          {"email"=>
-            "blah-foobar@developer.gserviceaccount.com"},
-        "organizer"=>
-          {"email"=>calendar_id,
-            "display_name"=>"Huginn Location Log",
-            "self"=>true},
-        "start"=>{"date_time"=>"2014-10-03T00:30:00+09:30"},
-        "end"=>{"date_time"=>"2014-10-03T01:30:00+09:30"},
-        "i_cal_uid"=>"blah@google.com",
-        "sequence"=>0,
-        "reminders"=>{"use_default"=>true}
-      }
+      { 'kind' => 'calendar#event',
+        'etag' => '"2908684044040000"',
+        'id' => 'baz',
+        'status' => 'confirmed',
+        'html_link' =>
+          'https://calendar.google.com/calendar/event?eid=foobar',
+        'created' => '2016-02-01T15:53:41.000Z',
+        'updated' => '2016-02-01T15:53:42.020Z',
+        'summary' => 'Awesome event',
+        'description' =>
+          'An example event with text. Pro tip: DateTimes are in RFC3339',
+        'creator' =>
+          { 'email' =>
+            'blah-foobar@developer.gserviceaccount.com' },
+        'organizer' =>
+          { 'email' => calendar_id,
+            'display_name' => 'Huginn Location Log',
+            'self' => true },
+        'start' => { 'date_time' => '2014-10-03T00:30:00+09:30' },
+        'end' => { 'date_time' => '2014-10-03T01:30:00+09:30' },
+        'i_cal_uid' => 'blah@google.com',
+        'sequence' => 0,
+        'reminders' => { 'use_default' => true } }
     end
 
     def setup_mock!
-      fake_interface = Object.new
-      expect(GoogleCalendar).to receive(:new).with(agent.interpolate_options(agent.options), Rails.logger) { fake_interface }
-      expect(fake_interface).to receive(:publish_as).with(calendar_id, message) { response_hash }
+      fake_interface = double('fake_interface')
+      expect(GoogleCalendar).to receive(:new).with(agent.interpolate_options(agent.options), Rails.logger) {
+                                  fake_interface
+                                }
+      allow(fake_interface).to receive(:publish_as).with(calendar_id, message) { response_hash }
       expect(fake_interface).to receive(:cleanup!)
     end
 
@@ -197,11 +201,12 @@ describe Agents::GoogleCalendarPublishAgent do
       it 'should publish old style payload it receives' do
         setup_mock!
 
-        expect {
+        expect do
           agent.receive([event])
-        }.to change { agent.events.count }.by(1)
+        end.to change { agent.events.count }.by(1)
 
-        expect(agent.events.last.payload).to eq({ "success" => true, "published_calendar_event" => response_hash, "agent_id" => event.agent_id, "event_id" => event.id })
+        expect(agent.events.last.payload).to eq({ 'success' => true, 'published_calendar_event' => response_hash,
+                                                  'agent_id' => event.agent_id, 'event_id' => event.id })
       end
     end
   end
