@@ -1,4 +1,4 @@
-ENV["RAILS_ENV"] ||= 'test'
+ENV['RAILS_ENV'] ||= 'test'
 
 if ENV['COVERAGE']
   require 'simplecov'
@@ -16,7 +16,8 @@ elsif ENV['CI'] == 'true'
   end
 end
 
-require File.expand_path('../config/environment', __dir__)
+# require File.expand_path('../config/environment', __dir__)
+require_relative '../config/environment'
 require 'rspec/rails'
 require 'webmock/rspec'
 
@@ -24,7 +25,7 @@ WebMock.disable_net_connect!(allow_localhost: true)
 
 # Requires supporting ruby files with custom matchers and macros, etc,
 # in spec/support/ and its subdirectories.
-Dir[Rails.root.join("spec/support/**/*.rb")].each { |f| require f }
+Dir[Rails.root.join('spec/support/**/*.rb')].each { |f| require f }
 
 ActiveRecord::Migration.maintain_test_schema!
 
@@ -61,27 +62,23 @@ RSpec.configure do |config|
   # to individual examples or groups you care about by tagging them with
   # `:focus` metadata. When nothing is tagged with `:focus`, all examples
   # get run.
-  if ENV['CI'] != 'true'
-    config.filter_run :focus
-  end
+  config.filter_run :focus if ENV['CI'] != 'true'
   config.run_all_when_everything_filtered = true
 
   # Run specs in random order to surface order dependencies. If you find an
   # order dependency and want to debug it, you can fix the order by providing
   # the seed, which is printed after each run.
   #     --seed 1234
-  config.order = "random"
+  config.order = 'random'
   config.global_fixtures = :all
 
   config.render_views
 
-  config.example_status_persistence_file_path = "./spec/examples.txt"
+  config.example_status_persistence_file_path = './spec/examples.txt'
 
   config.include Devise::Test::ControllerHelpers, type: :controller
   config.include SpecHelpers
   config.include ActiveSupport::Testing::TimeHelpers
 end
 
-if ENV['RSPEC_TASK'] != 'spec:nofeatures'
-  require 'capybara_helper'
-end
+require 'capybara_helper' if ENV['RSPEC_TASK'] != 'spec:nofeatures'
